reformat(utils): Move redirection sanitizer
